% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokyo.R
\name{altura}
\alias{altura}
\title{Tokio}
\usage{
altura(x, altura, sexo, s = NULL)
}
\arguments{
\item{x}{Un dataframe con informacion de personas. En el caso del ejemplo son deportistas olimpicos.}

\item{altura}{La variable que contiene la altura de cada persona.}

\item{sexo}{La variable que contiene el sexo de cada persona.}

\item{s}{Indicacion sobre el sexo de la persona que se desea filtrar.}
}
\value{
La o las persona de mayor altura del dataframe proporcionado.
}
\description{
Devuelve la persona con mayor altura. Ademas es posible filtrar por el sexo de
 la persona si lo considera necesario.
}
\examples{
nombre <- c("Juan", "Maria", "Ana","Pablo")
estatura <- c(182, 167, 173, 179)
sexo <- c("M","F", "F", "M")
atletas <- data.frame(nombre, estatura, sexo)
}
