% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atletas.R
\docType{data}
\name{atletas}
\alias{atletas}
\title{Data of nominal brown shrimp prices by size}
\format{
A Un DataFrame con cuatro filas y tres variables :
\describe{
  \item{nombre}{nombre del atleta}
  \item{estatura}{estatura del atleta en cm}
  \item{sexo}{sexo del atleta}
}
}
\source{
\url{Generados por el autor}
}
\usage{
atletas
}
\description{
Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound, etc. A smaller number per pound indicates larger shrimp. Nominal prices are total monthly value of brown shrimp andings within size class divided by total monthly landings within the size class.
}
\keyword{datasets}
